baseSize: Int --> baseSize: Long, map.plus --> map.put

+++ src/main/kotlin/DataBase.kt
@@ -1,4 +1,4 @@
 interface DataBase {
-    var baseSize : Int
+    var baseSize : Long
     // TODO: Add some more common properties to interface, or remove it
 }
